Issue #2: s_max should be removed
Replaced with s_softmax. There is no explicit constraint anymore that enforces that the total buffer memory remains below this threshold.
This value is still important, however, as it is used for scaling and to calculate the memory_load trajectories. The memory_load values can as a consequence have values > 1. This is taken into account in the cost function, where previously (1-memory_load)*(1-bandwidth_load) was used for scaling. This is now replaced with (1-bandwidth_load)/(1+memory_load).

Also adapted the most important files that make use of s_max.

ots_test.ipynb (tested)
ots_test.py (tested)